+++ leggedsnake/__init__.py
@@ -25,7 +25,7 @@ from pylinkage import (
 from .utility import step, stride
 from .walker import Walker
 from .geneticoptimizer import (
-    evolutionary_optimization, genetic_optimization
+    evolutionary_optimization, GeneticOptimization
 )
 from .dynamiclinkage import (
     Nail, PinUp, DynamicPivot, Motor,