+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- ``genetic_optimization`` class in ``geneticoptimizer.py`` that will replace the previous functional paradigm.
+- ``GeneticOptimization`` class in ``geneticoptimizer.py`` that will replace the previous functional paradigm.
 - ``VisualWorld`` has a new method called ``reload_visuals``.
 - ``show_evolution.py`` is a new script plotting various data about the Walkers population's evolution during 
 genetic optimization.